requirements.txt
# Web Framework
flask>=3.1.0,<4.0.0

# Web Scraping
selenium>=4.27.0,<5.0.0
beautifulsoup4>=4.12.3,<5.0.0
requests>=2.32.0,<3.0.0
google-search-results>=2.4.2,<3.0.0

# AI/LLM Integration
# LangChain ecosystem - all packages must be version-compatible
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-anthropic>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0

# LLM Provider SDKs
openai>=1.58.0,<2.0.0
anthropic>=0.40.0,<1.0.0

# Utilities
python-dotenv>=1.0.1,<2.0.0
urllib3>=2.2.3,<3.0.0

# AWS Deployment
boto3>=1.35.0,<2.0.0
gunicorn>=23.0.0,<24.0.0
